FIX: Removendo interface do Parcelamento service visto que não é necessário.
